Personalized Mobile Search Engine
The Project is based on data mining concept that captures the user's
preferences in the form of concepts by mining their click-through data.
The received SRR reduced on the basis of personal choice(browsing
history) and the location(GPS).
The user preferences are organized in an ontology-based, multifaceted
user profile, which are used to adapt a personalized ranking function
for rank adaptation of future search results.
The client collects and stores locally the click-through data to protect
privacy, whereas heavy tasks such as concept extraction, training, and
re-ranking are performed at the PMSE server .